requirements.txt
Babel==2.10.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
cvzone==1.5.6
easyocr==1.5.0
idna==3.3
imageio==2.19.3
imutils==0.5.4
networkx==2.8.3
numpy==1.22.4
opencv-python==4.5.4.60
opencv-python-headless==4.5.4.60
packaging==21.3
PIL-Tools==1.1.0
Pillow==9.1.1
pyparsing==3.0.9
python-bidi==0.4.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.1
six==1.16.0
tensorflow==1.15.5
tifffile==2022.5.4
tkcalendar==1.6.1
torch==1.11.0
torchvision==0.12.0
typing-extensions==4.2.0
urllib3==1.26.9